Develop a Main

+++ Parte 1 y 3/src/main/java/edu/eci/arst/concprg/prodcons/StartProduction.java
@@ -5,36 +5,24 @@
  */
 package edu.eci.arst.concprg.prodcons;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Queue;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 public class StartProduction {
-    
-    
+
     public static void main(String[] args) {
-        
-        Queue<Integer> queue=new LinkedBlockingQueue<>();
-        
-        
-        new Producer(queue,Long.MAX_VALUE).start();
-        
-        //let the producer create products for 5 seconds (stock).
+        Queue<Integer> queue = new LinkedBlockingQueue<>();
+
+        new Producer(queue, Integer.MAX_VALUE).start();
+
+        // let the producer create products for 5 seconds (stock).
         try {
             Thread.sleep(5000);
         } catch (InterruptedException ex) {
             Logger.getLogger(StartProduction.class.getName()).log(Level.SEVERE, null, ex);
         }
-        
-        
         new Consumer(queue).start();
     }
-    
-
 }
